Move InventoryMetricUnit enum to testAutomation to share between LKS and apps
