--- src/org/labkey/test/params/experiment/MetricUnit.java
@@ -1,7 +1,0 @@
-package org.labkey.test.params.experiment;
-
-public interface MetricUnit
-{
-    String getLabel(); // For UI
-    String getValue(); // For API
-}
